+++ packages/shared/scripts/generateCodeSamples.ts
@@ -42,6 +42,16 @@ const generateCodeSamples = (): void => {
       ],
     },
     {
+      component: 'p-modal',
+      samples: [
+        [
+          componentsJsPath + '/modal-example-accessibility.html',
+          componentsAngularPath + '/modal-example-accessibility.component.ts',
+          componentsReactPath + '/ModalExampleAccessibility.tsx',
+        ],
+      ],
+    },
+    {
       component: 'p-segmented-control',
       samples: [
         [
@@ -113,9 +123,14 @@ const generateCodeSamples = (): void => {
       component: 'p-tabs-bar',
       samples: [
         [
-          componentsJsPath + '/tabs-bar-example.html',
-          componentsAngularPath + '/tabs-bar-example.component.ts',
-          componentsReactPath + '/TabsBarExample.tsx',
+          componentsJsPath + '/tabs-bar-example-basic.html',
+          componentsAngularPath + '/tabs-bar-example-basic.component.ts',
+          componentsReactPath + '/TabsBarExampleBasic.tsx',
+        ],
+        [
+          componentsJsPath + '/tabs-bar-example-accessibility.html',
+          componentsAngularPath + '/tabs-bar-example-accessibility.component.ts',
+          componentsReactPath + '/TabsBarExampleAccessibility.tsx',
         ],
       ],
     },